# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChessCouncil",
    "description": "多 Agent 协作辩论的实时国际象棋分析与对战系统",
    "version": "0.4.1"
  },
  "paths": {
    "/api/game/new": {
      "post": {
        "summary": "New Game",
        "operationId": "new_game_api_game_new_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/move": {
      "post": {
        "summary": "Make Move",
        "operationId": "make_move_api_game_move_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/ai-step": {
      "post": {
        "summary": "Ai Step",
        "operationId": "ai_step_api_game_ai_step_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiStepRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/state": {
      "get": {
        "summary": "Get State",
        "operationId": "get_state_api_game_state_get",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/load-fen": {
      "post": {
        "summary": "Load Fen",
        "operationId": "load_fen_api_game_load_fen_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/analyze-position": {
      "post": {
        "summary": "Analyze Position",
        "description": "分析当前局面（不走子），用于 Demo / 识谱后 Council。",
        "operationId": "analyze_position_api_game_analyze_position_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalyzePositionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/game/undo": {
      "post": {
        "summary": "Undo Move",
        "description": "悔棋（人 vs AI 尽量回到人类回合）。",
        "operationId": "undo_move_api_game_undo_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library": {
      "get": {
        "summary": "Library List",
        "operationId": "library_list_api_library_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "game|endgame|puzzle",
              "title": "Category"
            },
            "description": "game|endgame|puzzle"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library/{item_id}/load": {
      "post": {
        "summary": "Library Load",
        "operationId": "library_load_api_library__item_id__load_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LibraryLoadRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/library/step": {
      "post": {
        "summary": "Library Step",
        "operationId": "library_step_api_library_step_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LibraryStepRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiStepRequest": {
        "properties": {
          "with_analysis": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Analysis"
          }
        },
        "type": "object",
        "title": "AiStepRequest"
      },
      "AnalyzePositionRequest": {
        "properties": {
          "with_analysis": {
            "type": "boolean",
            "title": "With Analysis",
            "default": true
          }
        },
        "type": "object",
        "title": "AnalyzePositionRequest"
      },
...
      "FenRequest": {
        "properties": {
          "fen": {
            "type": "string",
            "title": "Fen"
          }
        },
        "type": "object",
        "required": [
          "fen"
        ],
        "title": "FenRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryLoadRequest": {
        "properties": {
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "human_vs_human",
                  "human_vs_ai",
                  "ai_vs_ai"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "with_analysis": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Analysis"
          },
          "human_color": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "white",
                  "black"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Color"
          },
          "free_play": {
            "type": "boolean",
            "title": "Free Play",
            "default": false
          },
          "engine_depth": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 25.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Depth"
          }
        },
        "type": "object",
        "title": "LibraryLoadRequest"
      },
      "LibraryStepRequest": {
        "properties": {
          "with_analysis": {
            "type": "boolean",
            "title": "With Analysis",
            "default": false
          }
        },
        "type": "object",
        "title": "LibraryStepRequest"
      },
      "MoveRequest": {
        "properties": {
          "uci": {
            "type": "string",
            "title": "Uci"
          },
          "with_analysis": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Analysis"
          },
          "analysis_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fast",
                  "deep"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Mode"
          }
        },
        "type": "object",
        "required": [
          "uci"
        ],
        "title": "MoveRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}